Support scala 2.13.0-RC1
