+++ .travis.yml
@@ -2,6 +2,7 @@ language: scala
 scala:
 - 2.11.12
 - 2.12.6
+- 2.13.0-RC1
 sudo: false
 jdk:
 - oraclejdk8
